requirements.txt
python-dotenv==1.0.1
supabase==1.0.3
requests==2.32.3
matplotlib==3.10.0
openai==1.59.6
httpx==0.26.0